requirement.txt
django
django-restframework
django-cors-headers
djangorestframework-simplejwt